crime-pipeline/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
spacy==3.7.2
beautifulsoup4==4.12.3
requests==2.31.0
googlemaps==4.10.0
scikit-learn==1.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.2.0
boto3==1.34.34
